requirements.txt
Django==5.2.13
djangorestframework==3.14.0
django-cors-headers==4.0.0
